Multi-Agents/requirements.txt
python-dotenv
crewai[tools]
langchain
langchain-community
langchain-core